+++ product_code/configuration_service/impl/src/main/scala/de/upb/cs/uc4/configuration/impl/ConfigurationServiceImpl.scala
@@ -17,7 +17,7 @@ import play.api.Environment
 import scala.concurrent.duration._
 import scala.concurrent.{ ExecutionContext, Future }
 
-/** Implementation of the CourseService */
+/** Implementation of the ConfigurationService */
 class ConfigurationServiceImpl(override val environment: Environment)(implicit ec: ExecutionContext, config: Config)
   extends ConfigurationService {