+++ product_code/course_service/impl/src/main/scala/de/upb/cs/uc4/course/impl/CourseServiceImpl.scala
@@ -14,6 +14,7 @@ import de.upb.cs.uc4.course.impl.actor.CourseState
 import de.upb.cs.uc4.course.impl.commands._
 import de.upb.cs.uc4.course.impl.readside.CourseDatabase
 import de.upb.cs.uc4.course.model.Course
+import de.upb.cs.uc4.examreg.api.ExamregService
 import de.upb.cs.uc4.shared.client.exceptions._
 import de.upb.cs.uc4.shared.server.ServiceCallFactory._
 import de.upb.cs.uc4.shared.server.messages.{ Accepted, Confirmation, Rejected }
@@ -27,6 +28,7 @@ import scala.concurrent.{ Await, ExecutionContext, Future, TimeoutException }
 class CourseServiceImpl(
     clusterSharding: ClusterSharding,
     userService: UserService,
+    examregService: ExamregService,
     database: CourseDatabase,
     override val environment: Environment
 )(implicit ec: ExecutionContext, config: Config) extends CourseService {
@@ -40,7 +42,7 @@ class CourseServiceImpl(
   lazy val validationTimeout: FiniteDuration = config.getInt("uc4.timeouts.validation").milliseconds
 
   /** @inheritdoc */
-  override def getAllCourses(courseName: Option[String], lecturerId: Option[String]): ServerServiceCall[NotUsed, Seq[Course]] = authenticated[NotUsed, Seq[Course]](AuthenticationRole.All: _*) {
+  override def getAllCourses(courseName: Option[String], lecturerId: Option[String], moduleIds: Option[String]): ServiceCall[NotUsed, Seq[Course]] = authenticated[NotUsed, Seq[Course]](AuthenticationRole.All: _*) {
     ServerServiceCall { (header, _) =>
       database.getAll
         .map(seq => seq
@@ -51,13 +53,22 @@ class CourseServiceImpl(
             .filter(opt => opt.isDefined) //Filter every not existing course
             .map(opt => opt.get) //Future[Seq[Course]]
           ))
-        .map(seq => seq
-          .filter { course =>
-            //If courseName query is set, we check that every whitespace seperated parameter is contained
-            courseName.isEmpty || courseName.get.toLowerCase.split("""\s+""").forall(course.courseName.toLowerCase.contains(_))
-          }
-          .filter(course => lecturerId.isEmpty || course.lecturerId == lecturerId.get))
-        .map(courses => createETagHeader(header, courses))
+        .map { seq =>
+          seq
+            .filter { course =>
+              //If courseName query is set, we check that every whitespace seperated parameter is contained
+              courseName.isEmpty || courseName.get.toLowerCase.split("""\s+""").forall(course.courseName.toLowerCase.contains(_))
+            }
+            .filter(course => lecturerId.isEmpty || course.lecturerId == lecturerId.get)
+            .filter {
+              course =>
+                moduleIds match {
+                  case None => true
+                  case Some(listOfModuleIds) =>
+                    listOfModuleIds.split(',').exists(moduleId => course.moduleIds.contains(moduleId.trim))
+                }
+            }
+        }.map(courses => createETagHeader(header, courses))
     }
   }
 
@@ -71,12 +82,38 @@ class CourseServiceImpl(
             throw UC4Exception.OwnerMismatch
           }
 
-          val validationErrors = try {
+          var validationErrors = try {
             Await.result(courseProposal.validate, validationTimeout)
           }
           catch {
             case _: TimeoutException => throw UC4Exception.ValidationTimeout
             case e: Exception        => throw UC4Exception.InternalServerError("Validation Error", e.getMessage)
+          }
+
+          if (courseProposal.moduleIds.nonEmpty) {
+            val moduleCheckFuture = examregService.getModules(None, Some(true)).invoke().map {
+              modules =>
+                if (modules.isEmpty) {
+                  for (index <- courseProposal.moduleIds.indices) {
+                    validationErrors :+= SimpleError(s"moduleIds[$index]", "Module does not exist")
+                  }
+                }
+                else {
+                  val moduleIdList = modules.map(_.id)
+                  for (index <- courseProposal.moduleIds.indices) {
+                    if (!moduleIdList.contains(courseProposal.moduleIds(index)))
+                      validationErrors :+= SimpleError(s"moduleIds[$index]", "Module does not exist")
+                  }
+                }
+            }
+            try {
+              // TODO use another timeout variable, for internal server communication
+              Await.result(moduleCheckFuture, validationTimeout)
+            }
+            catch {
+              case _: TimeoutException => throw UC4Exception.ValidationTimeout
+              case e: Exception        => throw UC4Exception.InternalServerError("Validation Error", e.getMessage)
+            }
           }
 
           // If lecturerId is empty, the userService call cannot be found, therefore check and abort
@@ -106,6 +143,7 @@ class CourseServiceImpl(
                   throw UC4Exception(code, reason)
               }
           }
+
         }
     }
 
@@ -155,12 +193,38 @@ class CourseServiceImpl(
           throw UC4Exception.PathParameterMismatch
         }
 
-        val validationErrors = try {
+        var validationErrors = try {
           Await.result(updatedCourse.validate, validationTimeout)
         }
         catch {
           case _: TimeoutException => throw UC4Exception.ValidationTimeout
           case e: Exception        => throw UC4Exception.InternalServerError("Validation Error", e.getMessage)
+        }
+
+        if (updatedCourse.moduleIds.nonEmpty) {
+          val moduleCheckFuture = examregService.getModules(None, Some(true)).invoke().map {
+            modules =>
+              if (modules.isEmpty) {
+                for (index <- updatedCourse.moduleIds.indices) {
+                  validationErrors :+= SimpleError(s"moduleIds[$index]", "Module does not exist")
+                }
+              }
+              else {
+                val moduleIdList = modules.map(_.id)
+                for (index <- updatedCourse.moduleIds.indices) {
+                  if (!moduleIdList.contains(updatedCourse.moduleIds(index)))
+                    validationErrors :+= SimpleError(s"moduleIds[$index]", "Module does not exist")
+                }
+              }
+          }
+          try {
+            // TODO use another timeout variable, for internal server communication
+            Await.result(moduleCheckFuture, validationTimeout)
+          }
+          catch {
+            case _: TimeoutException => throw UC4Exception.ValidationTimeout
+            case e: Exception        => throw UC4Exception.InternalServerError("Validation Error", e.getMessage)
+          }
         }
 
         // If lecturerId is empty, the userService call cannot be found, therefore check and abort